AccesoRF/requirements.txt
functions-framework==3.*
requests
flask
firebase-admin
numpy
pytz
google-auth